[engine] simplify gltf files
