+++ src/Engine/Data/SpecularGlossinessMaterial.hpp
@@ -11,7 +11,15 @@ namespace Ra {
 namespace Engine {
 namespace Data {
 
-class SpecularGlossinessMaterialConverter;
+/**
+ * Radium IO to Engine conversion for pbrSpecularGlossiness
+ */
+class RA_ENGINE_API SpecularGlossinessMaterialConverter
+{
+  public:
+    Ra::Engine::Data::Material* operator()( const Ra::Core::Asset::MaterialData* toconvert );
+};
+
 /**
  * Radium Engine material representation of GLTF SpecularGlossiness Material
  * Texture semantics defined by this material :
